Replace several getters with read-only properties
This seems to be particularly valuable with regard to Plugin::config()
and ::version() which formerly were overloaded to work as getters and
setters. The usual naming convention would be clumsy for the getters
and ugly for the setters of the fluent interface.

And maybe we want to reintroduce ArrayAccess for Config and lang
later what had been removed with 67b08e7.