Substitute the routing with a more general mechanism
Formerly, the routing for user functions, plugin administration and
dynamic pages was handled individually. Furthermore it was not possible
to set controllers for requests, which didn't fall into one of these
categories.

Now we're able to attach controllers to arbitrary query string patterns,
even though we might have to be more verbose.

Furthermore we now assign the controllers explicitly, which avoids
potential security issues, makes it more clear what's going to happen,
and allows controllers to be reused more easily.

The resultant restriction on a simple dynamic dispatch of the actions
only (opposed to the double dispatch on controller and action) doesn't
seem to be a particular problem for most plugins, which have rather
limited functionality.