Switch to Doxygen as documentation generator
While we would really prefer sticking with PHPDocumentor, because its
written in PHP and, more importantly, caters better to some PHP specific
stuff, it is not up to the task yet. Besides some minor glitches, it is
not possible to include "external" documentation, and working around this
issue appears to be too clumsy for the readers.