Test controllers with Selenium instead of PHPUnit
It doesn't make sense to apply unit tests for controllers generally,
because too much has to be mocked, we wouldn't test important
details such as the actual language strings, and we wouldn't test
the real thing anyway.

Instead we opt in to Selenium test. For now, we stick with Selenese
tests which have to be run manually with SeleniumIDE. In the long run
we may consider to automate this.