Factor out class System
We want to gather the global state of the plugin framework in a central
place to avoid having several singletons and multitons, which appears
to be a general SRP violation. Particularly, removing the multiton
functionality out of Plugin seems to be valuable as that class still
does a lot (maybe too much).