Split the SDK into separate distribution
Contrary to 4c0575091144c73b89e24656a37fc81f321c4bcf, it really doesn't
make sense to include the pfw.phar and the API documentation in the
general plugin distribution. The slightly decreased convenience for
developers is outweighed by the much smaller plugin zip for almost all
users, and the added security to not have pfw.phar on a production
server.