Add View::__toString() for developer convenience
There are certainly cases where a view is not supposed to be echo'd, so
we offer a __toString() implementation that can be used to get the string
result. It is noteworthy, though, that any notices/warnings etc. are
generally suppressed in __toString(), so at least for testing, ::render()
should be preferred.