Close #3: Use "new style" push views
Being able to push closures to the view has disadvantages, namely that
the view is easily inclined to cache the results of calling some
closures, and even worse, that escaping by convention doesn't work in
all cases, for instance, when passing a closure result (which has
already been escaped) as additional argument to ::text(), which
absolutely has to escape its return value.

Therefore we're switching to pure push views, i.e. it's not allowed
anymore to push closures and call them from the view. This gives us the
option to to overload __call for view data (which triggers escaping),
but let __get access unescaped. Thus, a simple convention should be
sufficient to avoid XSS issues: only simple (i.e. non nested) method
calls on $this should be echo'd.

The drawback of having to construct sometimes very complex DTOs,
appears to be outweighed by the mentioned advantages. Furthermore, a
simple var_dump($this) in the view will show *all* available data.

Finally, we proclaim that views should only use a *very* limited subset
of PHP. It's not absolutely clear to us which exact subset, but for now
we want to use only `foreach` and `if` statements (besides implicit
`echo` via <?=), and variable, member access, dynamic method call and
`empty` expressions. Currently, we're also using string interpolation
which might not be the best idea.