Close #8: Add zero case to Lang::plural()
We want to express ourselves in proper language, so we add a special
case for zero items to ::plural(). We also change the suffixes to
numbers so that translators can more easily understand how to
properly localize. Furthermore, we remove the 2-4 fallback to 5,
because otherwise the proper translation might be easily overlooked.