Close #10: Add check for a required plugin
As plugins may depend on other plugins, we add a respective system
check. However, it is hard to check the version of *arbitrary* plugins
(at best we could parse version.nfo, if it exists), and the check might
be meaningless anyway, because it is not granted that a plugin follows
semantic versioning. Therefore we add a dedicated check for Pfw_XH
which allows to checks its version.