Close #16: Render human-readable form HTML
The HTML source code generated by `Form::render()` is hardly readable
for humans, because there are no line breaks and no indentation. This
is not so much a problem with regard to inspecting/debugging in a
browser, because contemporary developer tools make easily up for that.
However, the HTML is not necessarily viewed in a browser (for instance,
some test cases compare HTML strings), and therefore some pretty
printing would be appropriate. Luckily, the XMLWriter offers respective
settings.