Implement new view concept
This new view concept is based on our first stab in Pfw_XH 0.1,
and on the improved view concept which has later been implemented
in several CMSimple_XH plugins.  In all cases an instance of a
basically single `View` class has to be created, associated with a
template file and filled with relevant data by a controller (in
the broadest sense).  However, the way how the view data have to
be referenced has considerably changed; now all data are injected
into the template as local variables only (not as properties
and/or methods of the `View` object), and the data values are
automagically wrapped in `ViewValue` objects, which cater to
proper escaping when used in the template.

Supported PHP types for the view data are currently `bool`, `int`,
`float`, `string`, `array` and most objects, whereby `array` and
`Iterable` are actually only suitable for `foreach` loops, and
`if` clauses only handle `bool`s properly.  Other types can be
passed to the view, but don't appear to be to be useful, since it
is not possible to unwrap the underlying value of a `ViewValue`.
For instance, while passing a `resource` or a `callable` to the
view is possible, the template cannot access the actual values,
but only their string representation, which is meaningless for
such values.

Note that not only the values directly passed to the views are
wrapped as `ViewValue` objects, but also all values retrieved from
these objects.  For instance, looping over an `array` is supposed
to yield `ViewValue` objects encapsulating the actual array
values.  The basic idea is that it should be impossible to emit
invalid HTML from a template (unless a method called from the
template would have side effects, for instance `echo`ing
something).

To really make the views/templates "foolproof" with regard to
proper escaping (which is a prerequisite to prevent XSS attacks),
and also to keep them readable, the templates are supposed to use
only a minimal subset of the full PHP syntax.  Presently, we
consider the following syntax (expressed in EBNF) to be
appropriate:

````
template  = { html | php | print } ;
php       = "<?php" statement "?>" ;
print     = "<?=" expr "?>" ;
statement = foreach | "endforeach"
          | if | elseif | "else" ":" | "endif"
          | var "(" ")" ;
foreach   = "foreach" "(" expr "as" var "=>" var ")" ":" ;
if        = "if" "(" expr ")" ":" ;
elseif    = "elseif" "(" expr ")" ":" ;
expr      = var { "->" ident [ "(" [ expr { "," expr } ] ")" ] } ;
````

The non-terminal `var` stands for a PHP variable indentfier, and
the non-terminal `ident` for a PHP method identifier.

Note that the current view concept does not yet cater to
internationalization (besides PHP's `float` to `string`conversion,
which may not necessarily be desired, by the way), which has to be
handed in later.