Don't invoke nested view values
While that is certainly more efficient than converting nested views to
string and then echoing them, it yields harder to understand templates,
and might even undermine the automatic escaping.  Therefore we bite the
bullet and employ `__toString()` instead of `__invoke()`.

Actually, it is not clear whether there should even support for nested
view values, or whether these should be passed as `HtmlString`s in the
first place.  This issue deserves closer investigation.