Collapse View Hierarchy
It is not really clear if any other views than HTML views will ever be
needed.  If XML is to be created via a view it can be treated as HTML
view.  For plain text (such as email bodies) a view appears to be
overkill.  Therefore, we apply YAGNI, and collapse the `View`
hierarchy.