Close #24: Add phing target for minification of JS resources
We add a `minify` target which minifies all `.js` files in the plugin
root folder and creates respective `.min.js` files.  The `minify`
target is also called from the `build` target.  `.min.js` files of
plugins which already use the `build` target of Pfw_XH will not be
minified anymore.  Instead these plugins should be modified to have
`.js` files, and to include the `min.js` version if available, and to
fall back to the unminified `.js` version otherwise.

It is noteworthy, that the `build` target adds both `.js` and `.min.js`
files to the distribution, so that users could easily remove the
minified files for debugging and customization purposes.