Close #25: Abstract over the "test helpers"
We introduce our own TestCase base class, which offers the possibility
to mock functions and static methods, and to change the value of global
constants.  We keep it simple and use tcz/phpunit-mockfunction to
provide most of the needed functionality.  We also introduce some
factory methods which can be used to create instances *instead* of
using the `new` operator, which we cannot mock.